made edit route modal in profile, made profile more responsive, added dates to model
