+++ src/App.js
@@ -4,7 +4,6 @@ import Login from "./Pages/Login/Login";
 import FileUpload from "./Pages/FileUpload/FileUpload";
 import Home from "./Pages/Home/Home";
 import RunnerEditor from "./Pages/Editors/RunnerEditor";
-import ProfileCodeEditor from "./Pages/Editors/ProfileCodeEditor";
 import Peep from "./Pages/Peep/Peep";
 import Register from "./Pages/Register/Register.jsx";
 import Profile from "./Pages/Profile/Profile.jsx";
@@ -31,7 +30,6 @@ function App() {
           <Route path="/Home" exact component={Home} />
           <Route path="/Profile" exact component={Profile} />
           <Route path="/Editor" exact component={RunnerEditor} />
-          <Route path="/ProfileEditor" exact component={ProfileCodeEditor} />
           <Route path="/Peep" exact component={Peep} />
           <Route path="/Register" exact component={Register} />
         </Switch>
