--- src/Components/Modals/UpdateModale.jsx
@@ -1,56 +1,0 @@
-import React, { useState } from "react";
-import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from "reactstrap";
-
-const UpdateMod = (props) => {
-  const { buttonLabel, className } = props;
-
-  const [modal, setModal] = useState(false);
-
-  const toggle = () => setModal(!modal);
-
-  const externalCloseBtn = (
-    <button
-      className="close"
-      style={{ position: "absolute", top: "15px", right: "15px" }}
-      onClick={toggle}
-    >
-      &times;
-    </button>
-  );
-  return (
-    <div>
-      <Button className="p-3 mr-5 mt-4" color="primaryGrad" onClick={toggle}>
-        {buttonLabel}Update Profile
-      </Button>
-      <Modal
-        isOpen={modal}
-        toggle={toggle}
-        className={className}
-        external={externalCloseBtn}
-      >
-        <ModalHeader>Modal title</ModalHeader>
-        <ModalBody>
-          <b>Look at the top right of the page/viewport!</b>
-          <br />
-          Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
-          eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
-          minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-          aliquip ex ea commodo consequat. Duis aute irure dolor in
-          reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-          pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-          culpa qui officia deserunt mollit anim id est laborum.
-        </ModalBody>
-        <ModalFooter>
-          <Button color="primary" onClick={toggle}>
-            Do Something
-          </Button>{" "}
-          <Button color="secondary" onClick={toggle}>
-            Cancel
-          </Button>
-        </ModalFooter>
-      </Modal>
-    </div>
-  );
-};
-
-export default UpdateMod;
